requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asttokens==3.0.0
beautifulsoup4==4.12.3
biopandas==0.5.1
biopython==1.85
bioservices==1.11.2
biovec==0.2.7
cattrs==24.1.2
click==8.1.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cpdb-protein==0.2.0
cycler==0.12.1
cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
deepdiff==8.1.1
easydev==0.13.3
einops==0.8.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fair-esm==2.0.0
foldcomp==0.0.7
fonttools==4.55.3
gensim==3.4.0
gevent==24.11.1
graphein==1.7.7
greenlet==3.1.1
grequests==0.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jaxtyping==0.2.36
jedi==0.19.2
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.8
line-profiler==4.2.0
loguru==0.7.3
looseversion==1.1.2
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmtf-python==1.1.3
mpl-chord-diagram==0.3.2
msgpack==1.1.0
multipledispatch==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
orderly-set==5.2.3
packaging==24.2
pandas==2.2.3
pathlib2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prompt-toolkit==3.0.48
propy3==1.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyaaisc==1.2
pydantic==2.10.5
pydantic-core==2.27.2
pyfasta==0.5.2
pygments==2.19.1
pyprind==2.11.3
python-dateutil==2.9.0.post0
pytz==2024.2
torch-geometric==2.6.1
pyzmq==26.2.0
rdkit==2024.9.4
requests-cache==1.2.1
rich==13.9.4
rich-click==1.8.5
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selfies==2.2.0
six==1.17.0
smart-open==7.1.0
smilite==2.3.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
suds-community==1.2.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
# torch==2.5.1+cu124 # Need to be installed manually
# torchaudio==2.5.1+cu124
# torchvision==0.20.1+cu124
# torchsort
lightning
tensorboard
tornado==6.4.2
traitlets==5.14.3
tzdata==2024.2
url-normalize==1.4.3
wcwidth==0.2.13
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.1.1
xmltodict==0.14.2
zope-event==5.0
zope-interface==7.2
